Workaround in PHP cURL for the 100-continue expectation
cURL obeys the RFCs as it should. Meaning that for a HTTP/1.1 backend if the POST size is above 1024 bytes
cURL sends a 'Expect: 100-continue' header. The server acknowledges and sends back the '100' status code.
cuRL then sends the request body. This is proper behaviour. Nginx supports this header.

This allows to work around servers that do not support that header.
We're emptying the 'Expect' header, saying to the server: please accept the body right now.
Source: http://pilif.github.com/2007/02/the-return-of-except-100-continue/